Add Stream class, add combinators, improve tests
I think I prefer the do/transformer style over the applicative style.
The applicative style is actually quite difficult to support (and use)
due to nested functors. The transformer style is just a lot easier to
read and write. In fact it makes sense that transformers are better,
since what we have is nested monads, e.g. Parser s (Maybe a).

Also, I tried to support infinite streams, i.e. having a Stream that
always returns a token, but it's much nicer to just assume streams can
end. There's probably an abstraction that would make it nice (identity
functor?) but that's a lot of work to support something so niche.