Add new test using Either/ExceptT
We now have this !>> operator to say "or else this error", which uses
Either, which allows us to use ExceptT. So we give you the ability to
write parsers using MaybeT or ExceptT in a way that feels very similar
but allows different styles of error handling.

Parsing isn't too hard until you need to track positions and provide
good error messages. So this stuff is really important.

Removed weird nested functor operators and tests that use them. All hail
the do-block.

Added eof parser.

Generalised some parsers to Alternative.