Pull Maybe out of Stream class
Turns out this is totally fine, it just requires some monadic
combinators for convenience.

Notice that this:

  ExceptT $ eof !>> ExpectedEof

Became this:

  ExceptT $ match Nothing !>> ExpectedEof

In other words, eof is no longer a special case of stream handling -
it's just a regular stream value. And importantly, streams don't have to
have an eof value. We can easily define a stream that always returns a
token.