Add the zipper stream
This removes all the other instances for now, and adds (what I think
is) a very important type of Stream. Instead of iterating your
position, simply record the previous input. To get the position,
analyse the previous input. More importantly, you no longer need to
convert from an integer position to a printable line. Instead of
knowing the position, you have the location.