Rewrite parss!
* Adds `m` type parameter to Parser (standing for "match"). This is a
monoid that gets concatenated in the monad. In other words, composing
parsers also composes the matched text.

* We can write instances of Stream such as ([a], [a]), representing
the previous input and remaining input. This combines with the match
[a], to give us ([a], [a], [a]), representing the previous, matching
and remaining input. This enables excellent error messages.

* We provide convenience functions such as `locate` which
automatically applies your parse result to its location. In other
words, if you have a constructor where the last argument is a
location, simply return the constructor partially applied and `locate`
will do the rest.

* We provide 2 convenient monad transformer stacks: Infallible, which
adds Either to Parser, and Fallible which adds Either and Maybe to
Parser. This allows returning early with Nothing if the input is not
relevent or returning early with Left for an error variant.

* We provide convenience functions for operating in these stacks:
`infallible` to run the Infallible transformer, `fallible` to run
Fallible transformer, `ok` to lift, `need` to to potentially return
early and `fallback` to potentially return with an error variant.